Move back helpers into the watched folders
