+++ src/registrar/assets/modules/table-base.js
@@ -1,4 +1,4 @@
-import { hideElement, showElement, toggleCaret } from '../modules-common/helpers.js';
+import { hideElement, showElement, toggleCaret } from './helpers.js';
 
 export class LoadTableBase {
   constructor(sectionSelector) {
