+++ src/registrar/assets/modules/table-domain-requests.js
@@ -1,4 +1,4 @@
-import { hideElement, showElement, scrollToElement } from '../modules-common/helpers.js';
+import { hideElement, showElement, scrollToElement } from './helpers.js';
 import { initializeModals, unloadModals } from './helpers-uswds.js';
 import { getCsrfToken } from './helpers-csrf-token.js';
