+++ src/registrar/assets/modules/table-member-domains.js
@@ -1,6 +1,4 @@
-import { hideElement, showElement, scrollToElement, toggleCaret } from '../modules-common/helpers.js';
-import { initializeTooltips, initializeModals, unloadModals } from './helpers-uswds.js';
-import { getCsrfToken } from './helpers-csrf-token.js';
+import { scrollToElement } from './helpers.js';
 
 import { LoadTableBase } from './table-base.js';
